[05/02] style: Finalize getting random number
Using `/dev/urandom` is overkill, but a good practice for me.
It is cryptographically one of the most secure ways to get a random number.
Every operation of the system adds to the "entropy pool" the kernel uses.
Even if 2 processes start reading from it at the exact same nanosecond, the kernel will still serialize the reads.

https://web.archive.org/web/20200620004352/https://eternallyconfuzzled.com/using-rand-c-c-advice-for-the-c-standard-librarys-rand-function