[06] feat(utils): Add monadic functions to `Optional` and `Expected`
They are, however, much more limited in their use compared to the original ones.
The functions passed in the original ones can return different result or error types, but in my implementation the limits of C++98 and the lack of the `auto` keyword and `result_of` and `decltype` metaprogramming features make it very difficult to come close to that.