build: Exit with exit code `2` instead of `42` when build failed
From the GNU Make manual:
```
0
The exit status is zero if make is successful.

2
The exit status is two if make encounters any errors. It will print messages describing the particular errors.

1
The exit status is one if you use the ‘-q’ flag and make determines that some target is not already up to date.
```