feat(libftpp): Add `void_t`-like `voider` to `type_traits`
It is not named `void_t` because `::type` is still required to use it. The `_t` suffix is usually used for helper types that don't need `::type`.
The name `voider` is chosen because it is mentioned in N3911 (https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2014/n3911.pdf).
Interestingly, in the same document there's also a discussion that shows that the `void_t` name was contentious even for the standard library.