perf(libftpp): Change constant from `std::string` to char array
The constant gets constructed every function call, so `std::string` is a
bit heavy for that, given that it's a constant.

`static` was considered, but if it's within the function, it could lead
to many declarations since it's a template, and outside the function
encapsulation is violated and causes a declaration even if the template
never gets instantiated.